chore: rename wallet info library
