+++ pages/settings/Notifications.tsx
@@ -5,7 +5,7 @@ import { Label } from "~/components/ui/label";
 import { Switch } from "~/components/ui/switch";
 import { deregisterWalletNotifications } from "~/lib/notifications";
 import { useAppStore } from "~/lib/state/appStore";
-import { removeAllInfo } from "~/lib/storeWalletInfo";
+import { removeAllInfo } from "~/lib/walletInfo";
 import { registerForPushNotificationsAsync } from "~/services/Notifications";
 
 export function Notifications() {
